feat(transaction): set up a transaction client
