+++ libs/application/wallets/features/send-funds/src/lib/send-funds.component.ts
@@ -100,16 +100,10 @@ export class SendFundsComponent {
     this.submitted = true;
 
     if (this.sendFundsGroup.valid) {
-      this.transactionService
-        .transfer(
-          this.sendFundsGroup.get('recipient')?.value,
-          this.sendFundsGroup.get('amount')?.value
-        )
-        .subscribe({
-          next: () => console.log('emitted'),
-          complete: () => console.log('completed'),
-        });
-      //.subscribe(() => this.dialogRef.close());
+      this.transactionService.createTransaction(
+        this.sendFundsGroup.get('recipient')?.value,
+        this.sendFundsGroup.get('amount')?.value
+      );
     }
   }
 }
