+++ libs/solana-dapp/transaction/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/get-transfer-transaction';
 export * from './lib/types';
+export * from './lib/transaction-client';
