+++ libs/solana-dapp/wallet/src/lib/state/state.ts
@@ -1,8 +1,9 @@
 import { Network } from '@nx-dapp/solana-dapp/network';
+import { Transaction } from '@nx-dapp/solana-dapp/transaction';
 import { PublicKey } from '@solana/web3.js';
 import { SetNetworkAction } from '.';
 
-import { Wallet, WalletAdapter, WalletName, Transaction } from '../types';
+import { Wallet, WalletAdapter, WalletName } from '../types';
 import {
   ActionTypes,
   LoadWalletsAction,
